Move rendering logic to `MedicalHistorySummaryView`
